requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pyswisseph==2.10.3.2
geopy==2.4.1
timezonefinder==6.5.2
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
pytz==2024.1